set up primary-sidebar reusable component and refactor experiment + flowchart sidebar
