+++ src/components/experiment-tracking/experiment-primary-toolbar/index.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import PrimaryToolbar from '../../primary-toolbar';
+
+export const ExperimentPrimaryToolbar = ({
+  sidebarVisible,
+  setSidebarVisible,
+}) => {
+  return (
+    <PrimaryToolbar
+      visible={{ sidebar: sidebarVisible }}
+      onToggleSidebar={setSidebarVisible}
+    />
+  );
+};
+
+export default ExperimentPrimaryToolbar;
