--- src/components/experiment-tracking/sidebar/sidebar.scss
@@ -1,68 +1,0 @@
-@import '../../../styles/variables';
-@import '../../../styles/extends';
-
-.pipeline-sidebar {
-  @extend %sidebar;
-
-  position: absolute;
-  top: -1px; /* Avoids pixel rounding gaps */
-  bottom: -1px;
-  left: $global-toolbar-width;
-
-  // Ensures sidebar tooltips are above code panel
-  z-index: 1;
-  display: flex;
-  width: 100%;
-  transform: translateX(calc(-100% + #{$sidebar-width-closed}));
-  transition: transform ease 0.4s;
-
-  @media (min-width: $metadata-sidebar-code-breakpoint) {
-    z-index: 3;
-  }
-
-  @media (min-width: $sidebar-width-breakpoint) {
-    $sidebar-ui-width: $sidebar-width-open - $sidebar-width-closed;
-    width: $sidebar-width-open;
-    transform: translateX(-$sidebar-ui-width);
-  }
-
-  &--visible {
-    transform: translateX(0);
-  }
-
-  @media print {
-    display: none;
-  }
-
-  &--visible {
-    transform: translateX(0);
-  }
-}
-
-.pipeline-ui {
-  position: relative;
-  z-index: 2;
-  display: flex;
-  flex-direction: column;
-  flex-grow: 1;
-  height: 100%;
-  background: var(--color-bg-3);
-  visibility: hidden;
-  transition: visibility 0.3s;
-
-  .pipeline-sidebar--visible & {
-    visibility: visible;
-    transition: visibility 0s;
-  }
-}
-
-.pipeline-toolbar {
-  position: relative;
-  z-index: 2;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: $sidebar-width-closed;
-  height: 100%;
-  background: var(--color-bg-2);
-}
