+++ src/components/experiment-wrapper/index.js
@@ -11,6 +11,7 @@ import './experiment-wrapper.css';
 const MAX_NUMBER_COMPARISONS = 2; // 0-based, so three
 
 const ExperimentWrapper = ({ theme }) => {
+  const [sidebarVisible, setSidebarVisible] = useState(true);
   const [disableRunSelection, setDisableRunSelection] = useState(false);
   const [enableComparisonView, setEnableComparisonView] = useState(false);
   const [selectedRuns, setSelectedRuns] = useState([]);
@@ -80,9 +81,14 @@ const ExperimentWrapper = ({ theme }) => {
             onToggleComparisonView={onToggleComparisonView}
             runsListData={data.runsList}
             selectedRuns={selectedRuns}
+            sidebarVisible={sidebarVisible}
+            setSidebarVisible={setSidebarVisible}
           />
           {selectedRuns.length > 0 ? (
-            <Details selectedRuns={selectedRuns} />
+            <Details
+              selectedRuns={selectedRuns}
+              sidebarVisible={sidebarVisible}
+            />
           ) : null}
         </>
       ) : (
